Added Restaurants screen and info component
