+++ src/features/restaurants/screens/restaurants.screen.js
@@ -1,0 +1,31 @@
+import React from "react";
+import { View, SafeAreaView, StyleSheet, StatusBar } from "react-native";
+import { Searchbar } from "react-native-paper";
+import { RestaurantInfo } from "../components/restaurant-info.component";
+import { spacing } from "../../../utils/sizes";
+
+export const RestaurantsScreen = () => (
+  <SafeAreaView style={styles.container}>
+    <View style={styles.search}>
+      <Searchbar placeholder="Search" />
+    </View>
+    <View style={styles.list}>
+      <RestaurantInfo />
+    </View>
+  </SafeAreaView>
+);
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    marginTop: StatusBar.currentHeight,
+  },
+  search: {
+    padding: spacing.lg,
+  },
+  list: {
+    flex: 1,
+    padding: spacing.lg,
+    backgroundColor: "orange",
+  },
+});
